requirements.txt
slack_bolt
slack_sdk
python-dotenv
redis
requests
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
chromadb
openai
pydantic
pymongo
pytest>=7.4.0
pytest-asyncio>=0.21.0
